+++ Capstone/tenmo-server/src/main/java/com/techelevator/tenmo/dao/TransferDao.java
@@ -1,6 +1,5 @@
 package com.techelevator.tenmo.dao;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 import com.techelevator.tenmo.model.Transfer;
@@ -9,9 +8,9 @@ public interface TransferDao {
 
 	List<Transfer> viewTranserByUserAccountId(int userAccountId);
 	
-	Transfer sendMoney(int senderId, int recipientId, BigDecimal amountToTransfer);
+	Transfer sendMoney(int senderId, int recipientId, double amountToTransfer);
 	
-	Transfer requestMoney(int senderId, int personAskingId, BigDecimal amountToTransfer);
+	Transfer requestMoney(int senderId, int personAskingId, double amountToTransfer);
 	
 	List<Transfer> viewPending(int userAccountId);
